:abcd: New Lexer + Unit Tests
Writing a new lexer with goal of creating a new parser that uses this lexer as a base.

The lexer currently has some basic tokenization rules, and currently processes a whole block rather than a stream one character at a time. This eventually *may* need to be adapted when hooking it in to the connection stream, but for now simplifies the process a bit.

Included is a unit test suite for this code. These unit tests use Jest instead of the custom coded test script. I will be porting over those existing tests to Jest in the future, but for now they will remain intact.