:sparkle: Start Untagged Response Parsing
Starting the process of converting over UntaggedResponse parsing to the new parser.

Currently the parser is not hooked up to point to the new parsing code. that will be done after message-data parsing has been added, which is the last of the untagged response formats to add parsing for. However, it is also the most complex.

This change does, however, add in mailbox-data parsing, which includes listing, exists/recent counts, mailbox status, and search. Extended search is not yet supported.

Test plan:
Not hooked up, none yet