:name_badge: Make Untagged Type Overridable
Some untagged response types are easier to understand if they aren't a 1:1 match with the ATOM/Keyword. Most notably this is the case for `STATUS` (which would otherwise be `OK`/`NO`/`BAD`/etc.), and `MAILBOX-STATUS` (which would otherwise be `STATUS`, which is confusing with response status).

Adding support here just allows for a bit more user-friendly `type` value.